requirements.txt
pydantic>=2.0.0
pandas==2.3.0
numpy==2.2.6
PyYAML>=6.0
jsonschema>=4.0.0
httpx>=0.25.0
numba==0.61.2
python-dotenv>=1.0.0
redis>=5.0.0
APScheduler>=3.10.0

# tests requirements
pyarrow==20.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0